+++ src/vanilla/index.ts
@@ -1,0 +1,1 @@
+export { createStore } from './createStore'
